#6 change id's type in entities from Integer to Long

+++ src/main/java/com/example/booklibraryv2/repositories/LibraryUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibraryUserRepository extends JpaRepository<LibraryUser, Integer> {
+public interface LibraryUserRepository extends JpaRepository<LibraryUser, Long> {
 
   List<LibraryUser> findByNameContains(String searchQuery);
 }
